Refactor modal gallery rendering logic
Moved modal-specific gallery rendering and event handling from gallery.js to a new displayModalWorks function in modal.js. This improves separation of concerns and code maintainability by centralizing modal-related logic.
